python3/pyproject.toml
[tool.poetry]
name = "ase"
version = "0.1.0"
description = ""
authors = ["David Goade <dgoade@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
configparser = "^6.0.0"
pyyaml = "^6.0.1"
fabric = "^3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
